requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
anywhere==0.2.5
app==0.0.1
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
ddt==1.4.1
docopt==0.6.2
enum34==1.1.10
HTMLReport==2.1.5
idna==2.9
importlib-metadata==1.7.0
lxml==4.5.1
more-itertools==8.3.0
namedlist==1.7
numpy==1.18.5
packaging==20.4
pandas==1.0.4
parameterized==0.7.4
pipreqs==0.4.10
pitest==0.1
pluggy==0.13.1
py==1.8.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
pytest-check==0.3.9
pytest-html==2.1.1
pytest-metadata==1.9.0
pytest-rerunfailures==9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.4
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
yarg==0.1.9
zipp==3.1.0
